+++ napari_bioformats/__init__.py
@@ -3,6 +3,6 @@ try:
 except ImportError:
     __version__ = "unknown"
 
-from ._pims import napari_get_reader, read_bioformats
+from ._core import download_loci_jar, napari_get_reader, read_bioformats
 
-__all__ = ["napari_get_reader", "read_bioformats"]
+__all__ = ["napari_get_reader", "read_bioformats", "download_loci_jar"]
